move NewPipe functions to `NewPipeUtils`
