+++ innertube/src/main/java/com/zionhuang/innertube/InnerTube.kt
@@ -18,7 +18,6 @@ import io.ktor.serialization.kotlinx.json.*
 import io.ktor.util.encodeBase64
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.json.Json
-import org.schabi.newpipe.extractor.services.youtube.YoutubeJavaScriptPlayerManager
 import java.net.Proxy
 import java.util.*
 
@@ -141,7 +140,7 @@ class InnerTube {
                 playbackContext =
                     if (client.useSignatureTimestamp) {
                         PlayerBody.PlaybackContext(PlayerBody.PlaybackContext.ContentPlaybackContext(
-                            signatureTimestamp = YoutubeJavaScriptPlayerManager.getSignatureTimestamp(videoId)
+                            signatureTimestamp = NewPipeUtils.getSignatureTimestamp(videoId).getOrThrow()
                         ))
                     } else null
             )
